glob dependency is now only for testing
