+++ packages/foam-vscode/package.json
@@ -485,6 +485,7 @@
     "eslint-import-resolver-typescript": "^2.5.0",
     "eslint-plugin-import": "^2.24.2",
     "eslint-plugin-jest": "^25.3.0",
+    "glob": "^7.1.6",
     "husky": "^4.2.5",
     "jest": "^26.2.2",
     "jest-extended": "^0.11.5",
@@ -502,7 +503,6 @@
     "detect-newline": "^3.1.0",
     "fast-array-diff": "^1.0.1",
     "github-slugger": "^1.4.0",
-    "glob": "^7.1.6",
     "gray-matter": "^4.0.2",
     "lodash": "^4.17.21",
     "lru-cache": "^7.12.0",
