+++ packages/foam-vscode/src/core/janitor/generate-headings.test.ts
@@ -4,10 +4,10 @@ import { MarkdownResourceProvider } from '../services/markdown-provider';
 import { Resource } from '../model/note';
 import { Range } from '../model/range';
 import { FoamWorkspace } from '../model/workspace';
-import { FileDataStore } from '../services/datastore';
 import { Logger } from '../utils/log';
 import detectNewline from 'detect-newline';
 import { createMarkdownParser } from '../services/markdown-parser';
+import { FileDataStore } from '../../test/test-datastore';
 
 Logger.setLevel('error');
