+++ packages/foam-vscode/src/features/hover-provider.spec.ts
@@ -3,7 +3,6 @@ import { createMarkdownParser } from '../core/services/markdown-parser';
 import { MarkdownResourceProvider } from '../core/services/markdown-provider';
 import { FoamGraph } from '../core/model/graph';
 import { FoamWorkspace } from '../core/model/workspace';
-import { FileDataStore } from '../core/services/datastore';
 import {
   cleanWorkspace,
   closeEditors,
@@ -13,6 +12,7 @@ import {
 import { toVsCodeUri } from '../utils/vsc-utils';
 import { HoverProvider } from './hover-provider';
 import { readFileFromFs } from '../test/test-utils';
+import { FileDataStore } from '../test/test-datastore';
 
 // We can't use createTestWorkspace from /packages/foam-vscode/src/test/test-utils.ts
 // because we need a MarkdownResourceProvider with a real instance of FileDataStore.
